Add fork-awesome-stack.tsx and extract types
