+++ src/components/common/fork-awesome/fork-awesome-stack.tsx
@@ -1,0 +1,23 @@
+import React, { ReactElement } from 'react'
+import { ForkAwesomeIcon, ForkAwesomeIconProps } from './fork-awesome-icon'
+import { IconSize } from './types'
+
+export interface ForkAwesomeStackProps {
+  size?: IconSize
+  children: ReactElement<ForkAwesomeIconProps> | Array<ReactElement<ForkAwesomeIconProps>>
+}
+
+export const ForkAwesomeStack: React.FC<ForkAwesomeStackProps> = ({ size, children }) => {
+  return (
+    <span className={`fa-stack ${size ? 'fa-' : ''}${size ?? ''}`}>
+      {
+        React.Children.map(children, (child) => {
+          if (!React.isValidElement<ForkAwesomeIconProps>(child)) {
+            return null
+          }
+          return <ForkAwesomeIcon {...child.props} stacked={true}/>
+        })
+      }
+    </span>
+  )
+}
