+++ src/components/common/links/internal-link.tsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
-import { ForkAwesomeIcon, IconName } from '../fork-awesome/fork-awesome-icon'
+import { ForkAwesomeIcon } from '../fork-awesome/fork-awesome-icon'
+import { IconName } from '../fork-awesome/types'
 import { ShowIf } from '../show-if/show-if'
 import { LinkWithTextProps } from './types'
