+++ src/components/common/modals/common-modal.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import { Modal } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
-import { ForkAwesomeIcon, IconName } from '../fork-awesome/fork-awesome-icon'
+import { ForkAwesomeIcon } from '../fork-awesome/fork-awesome-icon'
+import { IconName } from '../fork-awesome/types'
 import { ShowIf } from '../show-if/show-if'
 
 export interface CommonModalProps {
